refactoring, move all code possible from bin/bit.js to app.ts, make it cleaner, readable and have all the bootstrap process in one file
